refactor(Money, Match, MatchPrizes, WinningNumbers, TicketCheckResult): move classes to domain package
