+++ src/main/java/lotto/domain/TicketCheckResult.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.dto.LottoResultItem;
 import lotto.dto.LottoWin;
