+++ src/test/java/LottoGameTest.java
@@ -1,7 +1,7 @@
-import lotto.WinningNumbers;
+import lotto.domain.WinningNumbers;
 import lotto.domain.LottoGame;
 import lotto.domain.LottoNumber;
-import lotto.Match;
+import lotto.domain.Match;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
