+++ src/test/java/MatchTest.java
@@ -1,4 +1,4 @@
-import lotto.Match;
+import lotto.domain.Match;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
