+++ src/test/java/MoneyTest.java
@@ -1,4 +1,4 @@
-import lotto.Money;
+import lotto.domain.Money;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
